+++ src/repo_analyzer/analyze.clj
@@ -76,8 +76,10 @@
 (defn compute-commit-time-distribution
           [commits]
           (->> commits
-               (map #(.format (java.text.SimpleDateFormat. "MM/dd/yyyy") (get-in % [:author :date] %)))
+               (map #(.format (java.text.SimpleDateFormat. "yyyy/MM/dd") (get-in % [:author :date] %)))
                frequencies
+               (map #(vector (first %) (second %)))
+               (sort-by (comp first))
                )
           )